reformat: edit view in create appointment
